auth_service/pyproject.toml

[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Authentication service"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"
httpx = "^0.27.0"
gunicorn = "^22.0.0"
orjson = "^3.10.7"
typer = "^0.12.3"
async-fastapi-jwt-auth = "^0.6.5"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
